+++ react/ModalHeader.tsx
@@ -13,6 +13,7 @@ import { useModalDispatch } from './components/ModalContext'
 interface Props {
   children?: React.ReactNode
   showCloseButton?: MaybeResponsiveInput<boolean>
+  iconCloseSize?: number
 }
 
 const CSS_HANDLES = [
@@ -22,8 +23,12 @@ const CSS_HANDLES = [
   'closeButtonContainer',
 ]
 
-export default function ModalHeader(props: Props) {
-  const { children, showCloseButton: showCloseButtonProp = true } = props
+export default React.memo(function ModalHeader(props: Props) {
+  const {
+    children,
+    iconCloseSize = 32,
+    showCloseButton: showCloseButtonProp = true,
+  } = props
 
   const showCloseButton = useResponsiveValue(showCloseButtonProp)
   const handles = useCssHandles(CSS_HANDLES)
@@ -41,7 +46,7 @@ export default function ModalHeader(props: Props) {
   const headerContainerClasses = classnames(
     styles.headerContainer,
     handles.headerContainer,
-    'flex items-start bb b--muted-3 flex-shrink-0 flex-grow-0',
+    'flex items-start bb b--muted-3 flex-shrink-0 flex-grow-0 z-5',
     {
       ['justify-between']: hasChildren,
       ['justify-end']: !hasChildren,
@@ -57,10 +62,10 @@ export default function ModalHeader(props: Props) {
             onClick={handleClose}
             className={`${handles.closeButton} ma0 bg-transparent pointer bw0 pa3`}
           >
-            <IconClose size={24} type="line" />
+            <IconClose size={iconCloseSize} type="line" />
           </button>
         </div>
       )}
     </div>
   )
-}
+})